// apps/foxxie/src/lib/Api/Spotify/MusicBrainzService.ts
export class MusicBrainzService {
    /**
     * MusicBrainz types of a group.
     */
    public groupTypes = ['Group', 'Orchestra'];

    /**
     * MusicBrainz types of a person.
     */
    public personTypes = ['Person', 'Violinist'];

    public async browseRecordingsForAnArtist(artist: string) {
        const musicBrainzRecordingsResult = await fetch(`https://musicbrainz.org/ws/2/recording`)
            .query({
                artist,
                limit: '2000',
                fmt: 'json'
            })
            .json<{ recordings: IRecording[] }>();
        return musicBrainzRecordingsResult;
    }

    public async fetchMusicBrainzData(artist: string) {
        const musicBrainzQueryResult = await fetch(`https://musicbrainz.org/ws/2/artist`)
            .query({
                query: artist,
                fmt: 'json'
            })
            .json<{ artists: { id: string; score: number }[] }>();

        const musicBrainzFoundArtist = musicBrainzQueryResult.artists.sort((a, b) => b.score - a.score)[0]?.id;
        if (!musicBrainzFoundArtist) return null;

        const musicBrainzResults = await fetch(`https://musicbrainz.org/ws/2/artist`)
            .path(musicBrainzFoundArtist)
            .query({
                inc: [
                    'recording-rels',
                    'work-rels',
                    'media',
                    'releases',
                    'various-artists',
                    'works',
                    'release-groups',
                    'recordings'
                ].join('+'),
                fmt: 'json'
            })
            .json<IArtist>();

        return musicBrainzResults;
    }

    public getInstrumentCredits(musicBrainzData: IArtist): InstrumentCredit[] {
        return (
            musicBrainzData.relations
                ?.filter(r => r['target-type'] === 'recording' && r.type === 'instrument' && r.direction === 'forward')
                .map(r => ({
                    type: r.attributes?.filter(a => a !== 'solo').join(', '),
                    name: r.recording.title,
                    link: `https://musicbrainz.org/recording/${r.recording.id}`
                })) || []
        );
    }
}

export interface InstrumentCredit {
    type?: string;
    name: string;
    link: string;
}

import DateTimeFormat = Intl.DateTimeFormat;

export interface IPeriod {
    begin: string;
    ended: boolean;
    end: string;
}

export interface IEntity {
    id: string;
}

export interface IArea extends IEntity {
    'iso-3166-1-codes': string[];
    name: string;
    'sort-name': string;
    disambiguation: string;
}

export interface IAlias extends IEntity {
    name: string;
    'sort-name': string;
    ended: boolean;
    'type-id': string;
    type: string;
    locale: string;
    primary: string;
    begin: string;
    end: string;
}

export interface IMatch {
    score: number; // ToDo: provide feedback: should be a number
}

export interface IArtist extends IEntity {
    name: string;
    disambiguation: string;
    'sort-name': string;
    'type-id'?: string;
    'gender-id'?: any;
    'life-span'?: IPeriod;
    country?: string;
    ipis?: any[]; // ToDo
    isnis?: string[];
    aliases?: IAlias[];
    gender?: string;
    type?: string;
    area?: IArea;
    begin_area?: IArea;
    end_area?: IArea;
    relations?: IRelation[];
    /**
     * Only defined if 'releases' are includes
     */
    releases?: IRelease[];
    'release-groups'?: IReleaseGroup[];
    recordings?: IRecording[];
}

export interface IArtistCredit {
    artist: IArtist;
    joinphrase: string;
    name: string;
}
// ...
export interface IRelease extends IEntity {
    title: string;
    'text-representation': { language: string; script: string };
    disambiguation: string;
    asin: string;
    'status-id': string;
    packaging?: string;
    status: string;
    'packaging-id'?: string;
    'release-events'?: IReleaseEvent[];
    date: string;
    media: IMedium[];
    'cover-art-archive': ICoverArtArchive;
    country: string;
    quality: string; // type ReleaseQuality doesnt work here
    barcode: string;
    relations?: IRelation[];
    'artist-credit'?: IArtistCredit[]; // Include 'artist-credits '
    'release-group'?: IReleaseGroup; // Include: 'release-groups'
}

export interface IReleaseEvent {
    area?: IArea;
    date?: string;
}

export type MediaFormatType = 'Digital Media'; // ToDo

export interface IRecording extends IEntity {
    video: boolean;
    length: number;
    title: string;
    disambiguation: string;
    isrcs?: string[];
    releases?: IRelease[];
    relations?: IRelation[];
    'artist-credit'?: IArtistCredit[];
    aliases?: IAlias[];
}

export interface ITrack extends IEntity {
    position: number;
    recording: IRecording;
    number: string; // in JSON, this is a string field
    length: number;
    title: string;
    'artist-credit'?: IArtistCredit[];
}

export interface IMedium {
    title: string;
    format?: string; // optional, type dosent work
    'format-id': string;
    tracks: ITrack[];
    'track-count': number;
    'track-offset': number;
    position: number;
}

export interface ICoverArtArchive {
    count: number;
    front: boolean;
    darkened: boolean;
    artwork: boolean;
    back: boolean;
}

export interface IReleaseGroup extends IEntity {
    count: number;
    disambiguation?: string;
    title: string;
    'secondary-types': string[];
    'first-release-date': string;
    'primary-type': string;
    'primary-type-id'?: string;
    'secondary-type-ids'?: string[];
    'sort-name': string;
    'artist-credit': { artist: IArtist; name: string; joinphrase: string }[];
    releases?: IRelease[]; // include 'releases'
}
// ...
export type RelationDirection = 'backward' | 'forward';

export interface IRelation {
    'attribute-ids': any;
    direction: RelationDirection;
    'target-credit': string;
    end: null | unknown;
    'source-credit': string;
    ended: boolean;
    'attribute-values': unknown;
    attributes?: any[];
    type: string;
    begin?: null | unknown;
    'target-type'?: 'url' | 'recording';
    'type-id': string;
    url?: IUrl;
    release?: IRelease;
    recording: IRecording;
}

export interface IRelationList {
    relations: IRelation[];
}

export interface IWork extends IEntity {
    title: string;
}

export interface ILabel extends IEntity {
    name: string;
}

export interface IPlace extends IEntity {
    name: string;
}

export interface IUrl extends IEntity {
    id: string;
    resource: string;
    'relation-list'?: IRelationList[];
}